reverted the fix for cyclic imports - not needed anymore
