+++ api/groups/nodeGroup_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/ElrondNetwork/elrond-go/config"
 	"github.com/ElrondNetwork/elrond-go/debug"
 	"github.com/ElrondNetwork/elrond-go/heartbeat/data"
+	"github.com/ElrondNetwork/elrond-go/node/external"
 	"github.com/ElrondNetwork/elrond-go/statusHandler"
 	"github.com/ElrondNetwork/elrond-go/testscommon"
 	"github.com/gin-gonic/gin"
@@ -136,7 +137,7 @@ func TestP2PMetrics_ShouldReturnErrorIfFacadeReturnsError(t *testing.T) {
 	expectedErr := errors.New("i am an error")
 
 	facade := mock.FacadeStub{
-		StatusMetricsHandler: func() statusHandler.StatusMetricsHandler {
+		StatusMetricsHandler: func() external.StatusMetricsHandler {
 			return &testscommon.StatusMetricsStub{
 				StatusP2pMetricsMapCalled: func() (map[string]interface{}, error) {
 					return nil, expectedErr
@@ -166,7 +167,7 @@ func TestNodeStatus_ShouldReturnErrorIfFacadeReturnsError(t *testing.T) {
 	expectedErr := errors.New("i am an error")
 
 	facade := mock.FacadeStub{
-		StatusMetricsHandler: func() statusHandler.StatusMetricsHandler {
+		StatusMetricsHandler: func() external.StatusMetricsHandler {
 			return &testscommon.StatusMetricsStub{
 				StatusMetricsMapWithoutP2PCalled: func() (map[string]interface{}, error) {
 					return nil, expectedErr
@@ -202,7 +203,7 @@ func TestStatusMetrics_ShouldDisplayNonP2pMetrics(t *testing.T) {
 	statusMetricsProvider.SetStringValue(p2pKey, "p2p value")
 
 	facade := mock.FacadeStub{}
-	facade.StatusMetricsHandler = func() statusHandler.StatusMetricsHandler {
+	facade.StatusMetricsHandler = func() external.StatusMetricsHandler {
 		return statusMetricsProvider
 	}
 
@@ -235,7 +236,7 @@ func TestP2PStatusMetrics_ShouldDisplayNonP2pMetrics(t *testing.T) {
 	statusMetricsProvider.SetStringValue(p2pKey, p2pValue)
 
 	facade := mock.FacadeStub{}
-	facade.StatusMetricsHandler = func() statusHandler.StatusMetricsHandler {
+	facade.StatusMetricsHandler = func() external.StatusMetricsHandler {
 		return statusMetricsProvider
 	}
 
@@ -458,7 +459,7 @@ func TestPrometheusMetrics_ShouldReturnErrorIfFacadeReturnsError(t *testing.T) {
 	expectedErr := errors.New("i am an error")
 
 	facade := mock.FacadeStub{
-		StatusMetricsHandler: func() statusHandler.StatusMetricsHandler {
+		StatusMetricsHandler: func() external.StatusMetricsHandler {
 			return &testscommon.StatusMetricsStub{
 				StatusMetricsWithoutP2PPrometheusStringCalled: func() (string, error) {
 					return "", expectedErr
@@ -491,7 +492,7 @@ func TestPrometheusMetrics_ShouldWork(t *testing.T) {
 	statusMetricsProvider.SetUInt64Value(key, value)
 
 	facade := mock.FacadeStub{}
-	facade.StatusMetricsHandler = func() statusHandler.StatusMetricsHandler {
+	facade.StatusMetricsHandler = func() external.StatusMetricsHandler {
 		return statusMetricsProvider
 	}
