+++ cmd/node/factory/structsStatusHandlers.go
@@ -8,6 +8,7 @@ import (
 	"github.com/ElrondNetwork/elrond-go-core/data/typeConverters"
 	"github.com/ElrondNetwork/elrond-go-core/marshal"
 	logger "github.com/ElrondNetwork/elrond-go-logger"
+	"github.com/ElrondNetwork/elrond-go/node/external"
 	"github.com/ElrondNetwork/elrond-go/statusHandler"
 	"github.com/ElrondNetwork/elrond-go/statusHandler/persister"
 	"github.com/ElrondNetwork/elrond-go/storage"
@@ -18,7 +19,7 @@ var log = logger.GetOrCreate("main")
 // StatusHandlersInfo is struct that stores all components that are returned when status handlers are created
 type statusHandlersInfo struct {
 	AppStatusHandler  core.AppStatusHandler
-	StatusMetrics     statusHandler.StatusMetricsHandler
+	StatusMetrics     external.StatusMetricsHandler
 	PersistentHandler *persister.PersistentStatusHandler
 }
 
@@ -90,7 +91,7 @@ func (shi *statusHandlersInfo) StatusHandler() core.AppStatusHandler {
 }
 
 // Metrics returns the status metrics
-func (shi *statusHandlersInfo) Metrics() statusHandler.StatusMetricsHandler {
+func (shi *statusHandlersInfo) Metrics() external.StatusMetricsHandler {
 	return shi.StatusMetrics
 }
