+++ testscommon/statusHandler/statusHandlersUtilsMock.go
@@ -2,13 +2,13 @@ package statusHandler
 
 import (
 	"github.com/ElrondNetwork/elrond-go-core/core"
-	"github.com/ElrondNetwork/elrond-go/statusHandler"
+	"github.com/ElrondNetwork/elrond-go/node/external"
 	"github.com/ElrondNetwork/elrond-go/storage"
 )
 
 // StatusHandlersUtilsMock -
 type StatusHandlersUtilsMock struct {
-	StatusMetrics    statusHandler.StatusMetricsHandler
+	StatusMetrics    external.StatusMetricsHandler
 	AppStatusHandler core.AppStatusHandler
 }
 
@@ -23,7 +23,7 @@ func (shum *StatusHandlersUtilsMock) StatusHandler() core.AppStatusHandler {
 }
 
 // Metrics -
-func (shum *StatusHandlersUtilsMock) Metrics() statusHandler.StatusMetricsHandler {
+func (shum *StatusHandlersUtilsMock) Metrics() external.StatusMetricsHandler {
 	return shum.StatusMetrics
 }
